requirements.txt
altgraph==0.17.3
async-timeout==4.0.2
azure-cognitiveservices-speech==1.27.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
Flask==2.2.3
future==0.18.3
idna==3.4
install==1.3.5
iso8601==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
MouseInfo==0.1.3
pefile==2023.2.7
Pillow==9.5.0
PyAudio==0.2.13
PyAutoGUI==0.9.53
pycparser==2.21
pyFirmata==1.1.0
PyGetWindow==0.0.9
pyinstaller==5.10.1
pyinstaller-hooks-contrib==2023.2
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.28
pyserial==3.5
pyttsx3==2.90
pytweening==1.0.4
pywhatkit==5.4
pywin32==306
pywin32-ctypes==0.2.0
PyYAML==6.0
redis==4.5.4
requests==2.28.2
serial==0.0.97
six==1.16.0
sounddevice==0.4.6
soupsieve==2.4
SpeechRecognition==3.10.0
urllib3==1.26.15
Werkzeug==2.2.3
wikipedia==1.4.0